# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Food Facts API",
        "description": "API com informações nutricionais de alimentos",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/tokens/create": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Criação de token",
                "description": "Criação de token de usuário para autênticação no uso da API",
                "operationId": "createToken",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "email",
                                        "example": "user@local"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    },
                                    "token_name": {
                                        "type": "string",
                                        "example": "para_teste"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "plainTextToken": {
                                            "type": "string",
                                            "example": "1|lQ4sE0RwgWzHCRCKsV13ISg6jZlwWb2F9kOlNqCa627d5fd9"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "422": {
                        "description": "Credênciais inválidas"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "API Token",
                "scheme": "bearer"
            }
        }
    },
...
}